+++ lib/output/writer/elasticsearch_test.go
@@ -107,12 +107,16 @@ func TestElasticIntegration(t *testing.T) {
 	t.Run("TestElasticConnect", func(te *testing.T) {
 		testElasticConnect(urls, client, te)
 	})
+
+	t.Run("TestElasticIndexInterpolation", func(te *testing.T) {
+		testElasticIndexInterpolation(urls, client, te)
+	})
 }
 
 func testElasticConnect(urls []string, client *elastic.Client, t *testing.T) {
 	conf := NewElasticsearchConfig()
 	conf.Index = "test_conn_index"
-	conf.ID = "${!count:foo}"
+	conf.ID = "foo-${!count:foo}"
 	conf.URLs = urls
 
 	m, err := NewElasticsearch(conf, log.New(os.Stdout, log.Config{LogLevel: "NONE"}), metrics.DudType{})
@@ -145,7 +149,70 @@ func testElasticConnect(urls []string, client *elastic.Client, t *testing.T) {
 		}
 	}
 	for i := 0; i < N; i++ {
-		id := fmt.Sprintf("%v", i+1)
+		id := fmt.Sprintf("foo-%v", i+1)
+		get, err := client.Get().
+			Index("test_conn_index").
+			Type("doc").
+			Id(id).
+			Do(context.Background())
+		if err != nil {
+			t.Fatalf("Failed to get doc '%v': %v", id, err)
+		}
+		if !get.Found {
+			t.Errorf("document %v not found", i)
+		}
+
+		var sourceBytes []byte
+		sourceBytes, err = get.Source.MarshalJSON()
+		if err != nil {
+			t.Error(err)
+		} else {
+			if exp, act := string(testMsgs[i][0]), string(sourceBytes); exp != act {
+				t.Errorf("wrong user field returned: %v != %v", act, exp)
+			}
+		}
+	}
+}
+
+func testElasticIndexInterpolation(urls []string, client *elastic.Client, t *testing.T) {
+	conf := NewElasticsearchConfig()
+	conf.Index = "${!metadata:index}"
+	conf.ID = "bar-${!count:bar}"
+	conf.URLs = urls
+
+	m, err := NewElasticsearch(conf, log.New(os.Stdout, log.Config{LogLevel: "NONE"}), metrics.DudType{})
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	if err = m.Connect(); err != nil {
+		t.Fatal(err)
+	}
+
+	defer func() {
+		m.CloseAsync()
+		if cErr := m.WaitForClose(time.Second); cErr != nil {
+			t.Error(cErr)
+		}
+	}()
+
+	N := 10
+
+	testMsgs := [][][]byte{}
+	for i := 0; i < N; i++ {
+		testMsgs = append(testMsgs, [][]byte{
+			[]byte(fmt.Sprintf(`{"user":"%v","message":"hello world"}`, i)),
+		})
+	}
+	for i := 0; i < N; i++ {
+		msg := message.New(testMsgs[i])
+		msg.Get(0).Metadata().Set("index", "test_conn_index")
+		if err = m.Write(msg); err != nil {
+			t.Fatal(err)
+		}
+	}
+	for i := 0; i < N; i++ {
+		id := fmt.Sprintf("bar-%v", i+1)
 		get, err := client.Get().
 			Index("test_conn_index").
 			Type("doc").